requirements.txt
awsebcli==3.9.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.3
botocore==1.4.91
cement==2.8.2
cffi==1.9.1
colorama==0.3.7
cryptography==1.7.2
Django==1.9
django-blog-zinnia==0.18.1
django-contrib-comments==1.8.0
django-coverage==1.2.4
django-mptt==0.8.7
django-tagging==0.4.5
django-xmlrpc==0.1.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
enum34==1.1.6
idna==2.2
ipaddress==1.0.18
jmespath==0.9.0
mots-vides==2015.5.11
pathspec==0.5.0
Pillow==3.4.2
pyasn1==0.1.9
pycparser==2.17
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
regex==2017.2.8
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
websocket-client==0.40.0
